==> 02_sklearn_multiclass.ipynb <==
Most of this should look very familiar from our last notebook. We're just increasing the difficulty by training a model to identify each of our 40 classes, rather than just two. We also introduce a new preprocesses step - removing stopwords.

import pandas as pd
import numpy as np
import sklearn

from google.colab import drive

drive.mount('/content/gdrive')

train = pd.read_csv('gdrive/My Drive/RTANews_raw/arabic_train.csv')
val = pd.read_csv('gdrive/My Drive/RTANews_raw/arabic_val.csv')
test = pd.read_csv('gdrive/My Drive/RTANews_raw/arabic_test.csv')

train.head()

#Due to RAM issues on Google Colab, we'll focus on the first 20 classes in our data.
train = train[train.label <= 20]
test = test[test.label <= 20]

In the last notebook, we passed all of our text into our model. But what about simple words like prepositions, that appear very often and don't necessarily contain helpful information about the meaning of the text?

Those simple, often repeated words are often called stopwords. A common preprocessing step is to remove those words from the text.

We won't be exploring it here, but `TfidfVectorizer` can also help address this problem, by essential penalizing words that appear very often -- including words that might not be stopwords. However, even in this case you'd probably want to remove stopwords.

#we'll use nltk to download and then load in a set of Arabic stopwords.
import nltk

nltk.download('stopwords')
stop_words = nltk.corpus.stopwords.words('arabic')

stop_words[0:5]

#Removing stopwords is easy - just an additional argument when we create our CountVectorizer!
vectorizer = sklearn.feature_extraction.text.CountVectorizer(ngram_range=(1,3), stop_words=stop_words)

words_train = [text for text in train.text]
words_test = [text for text in test.text]

X_train = vectorizer.fit_transform(words_train)
X_test = vectorizer.transform(words_test)

Y_train = train.label
Y_test = test.label

Conveniently, most `sklearn` classifiers are set up to do multiclass classification by default. You just feed in your data as you would for binary classsification and add one argument to specify that this is a multiclass problem. If you're interested in details on this, and the full list of classifiers that are inherently multiclass, see: https://scikit-learn.org/stable/modules/multiclass.html.

The next two cells will take a while to run.

from sklearn.linear_model import LogisticRegression

classifier = LogisticRegression(multi_class='multinomial').fit(X_train, Y_train)
classifier.score(X_test, Y_test)

First, the "STOP: TOTAL NO. of ITERATIONS REACHED LIMIT" message means that in the default number of iterations (1000), the model has failed to converge. We could address this by increased the number of iterations but for now we'll just move forward.

Unsurprisingly, our results aren't as good as they were for binary classification. But we're also not using the best metric here! Accuracy hides a lot of important information for multiclass problems. Instead, we'll calculate the F1 score, a better metric that takes into account both precision and recall (more here: https://towardsdatascience.com/accuracy-precision-recall-or-f1-331fb37c5cb9).

We'll look at the confusion matrix again. In this case `pd.crosstab` works better visually than `sklearn`'s confusion matrix.

from sklearn.metrics import f1_score

preds = classifier.predict(X_test)
f1_score = f1_score(Y_test, preds, average = 'weighted')
print(f'F1 score:{f1_score}\n\n')

pd.crosstab(Y_test, preds)


Here, we can see visually that the model is performing fairly well, and identify a few areas where it's failing. For example, categories 6 and 12 are most often confused.

print(train.category[train.label == 6].unique(), train.category[train.label == 12].unique())

Great, this makes sense! Category 6 is "Syrian opposition" and category 12 is "Syrian conflict." Without knowing the guidelines provided to the labelers, even a human might have a hard time dividing articles accurately into these two categories!
