==> 03_word2vec.ipynb <==
Approaches to representing text like `CountVectorizer` are sometimes called bag-of-words approaches, because they don't contain any information about the sequencing of the words. We're also using a fairly small dataset, so our model doesn't necessarily learn any general information about the mea

The next few notebooks will introduce more advanced techniques that try to address this problem. First, we'll take a look at word embeddings by focusing on the implementation of Word2Vec available in the `gensim` library.

import gensim
import pandas as pd
import numpy as np
import sklearn

from google.colab import drive

drive.mount('/content/gdrive')

train = pd.read_csv('gdrive/My Drive/RTANews_raw/arabic_train.csv')
val = pd.read_csv('gdrive/My Drive/RTANews_raw/arabic_val.csv')
test = pd.read_csv('gdrive/My Drive/RTANews_raw/arabic_test.csv')

train.head()

#Again, we'll limit ourselves to 20 classes for now.
train = train[train.label <= 20]
test = test[test.label <= 20]

We need to add a new preprocessing step here - tokenizing our text into individual words. `nltk`, which we used in a previous notebook for loading stopwords, has a nice utility for this. We're using a naive tokenizer that just looks for spaces between words, but `nltk` offers more advanced approaches as well, for example making use of regular expressions.

We'll also remove stopwords again.

import nltk

nltk.download('stopwords')
stop_words = nltk.corpus.stopwords.words('arabic')

from nltk.tokenize import WhitespaceTokenizer

tokenizer = WhitespaceTokenizer()

train_words = [tokenizer.tokenize(t) for t in train.text]
test_words = [tokenizer.tokenize(t) for t in test.text]

train_words = [[t for t in text if t not in stop_words] for text in train_words]
test_words = [[t for t in text if t not in stop_words] for text in test_words]

Now we're going to train a set of word embeddings using the popular Word2Vec approach.

Word embeddings are high-dimensional vector representations of words. In previous notebooks, we were also providing vector representations of our text, but using sparse vectors, or vectors where most values are zero ("this word appears zero times in this document") and therefore don't contain any information.

Word2Vec is a specific approach for learning word embeddings with a shallow neural network. You can read more about it, and word embeddings in general, here: https://towardsdatascience.com/introduction-to-word-embedding-and-word2vec-652d0c2060fa.

We pass a few arguments when creating our model, to define the minimum number of times a word can appear to be included in the model and to use the skigram implementation of word2vec (the other being cbow, or continous bag of words, which does better for datasets larger than ours).

model = gensim.models.Word2Vec(train_words, min_count=1, sg=1)

Now we should be able to get the embedding for an indiviudal word out of our trained model. Let's give it a try:

model['المعارضة']

A brief tangent: one appealing characteristic of word embeddings is that similar words should be near to each other in embedding space. `gensim` allows us to test this by computing the similarity between two words and by returning the most similar words to a given word.

So let's give it a try!

model.similarity('الحرب', 'الصراع')

So the words 'war' and 'conflict' are fairly similar - makes sense.

Now, we'll look at the five words most similar to "opposition." This is also built into the `gensim` implementation of word2vec.

model.most_similar('المعارضة')[:5]

This looks decent! Two of these are variants on the word opposition and the others are "armed," "factions," and "moderate" -- all words that Arab media outlets might use to describe an opposition movement.

We could further improve our results here by using a different metric, cosine similarity, to measure the distance between words in embedding space. But for now, let's return to our classification task.

Right now, we have vectors for individual words, but we want a single vector for each piece of text in our dataset. One way to solve this problem is by averaging the vectors of individual words. Let's create a simple function that does this.

#Remember that we've already tokenized our text, creating a 'list of lists.' Let's plan to feed that into our function.
def doc_vectorizer(text, model):
  doc_vec = 0
  count = 0

  if len(text) == 1:
    return model[text]

  for t in text:
    try:
      word_vec = model[t]
      doc_vec = doc_vec + word_vec
      count += 1
    except:
      pass
  
  doc_vec = doc_vec / count
  return doc_vec

X_train = [doc_vectorizer(t, model) for t in train_words]
X_test = [doc_vectorizer(t, model) for t in test_words]

For similicity and for better comparison, let's feed our new vectors into the same `sklearn` classifier that we've used previously.

from sklearn.linear_model import LogisticRegression

Y_train = train.label
Y_test = test.label

classifier = LogisticRegression(max_iter = 5000, multi_class='multinomial').fit(X_train, Y_train)
preds = classifier.predict(X_test)
pd.crosstab(Y_test, preds)

from sklearn.metrics import f1_score

f1_score(Y_test, preds, average = 'weighted')

This looks good. Notice that the model is getting confused between categories 6 and 12 less often. The F1 score is also slightly improved.

But wait! There's another approach we can try here. `gensim` also offers a doc2vec implementation, for cases like ours where we really want document-level and not a word-level vectors. We use a different function within `gensim` to tokenize our documents.

from gensim.models.doc2vec import Doc2Vec, TaggedDocument

documents = [TaggedDocument(doc, [i]) for i, doc in enumerate(train.text)]
doc_model = Doc2Vec(documents, min_count=1)

X_train = [doc_model.infer_vector(t) for t in train.text]
X_test = [doc_model.infer_vector(t) for t in test.text]


classifier = LogisticRegression(max_iter = 5000, multi_class='multinomial').fit(X_train, Y_train)
preds = classifier.predict(X_test)
f1_score(Y_test, preds, average = 'weighted')

These results are much worse! That's probably due to the size of our training set. With more data, doc2vec would likely outperform word2vec for this kind of problem.

In the next notebook, we'll explore how to take advantage of large training sets when we don't have access to them ourselves.
